+++ skills/menu.py
@@ -37,17 +37,3 @@ def get_menu(message):
             ]
         )
     )
-
-def get_lunch(message):
-    return TemplateSendMessage(
-        alt_text='열린 점심 위치',
-        template=ButtonsTemplate(
-            actions=[
-                MessageAction(label='장소', text='피알앤디컴퍼니'),
-                MessageAction(label='ML2', text='ML2'),
-                MessageAction(label='미래에셋대우', text='미래에셋대우'),
-                MessageAction(label='소프트웨어마이스터고등학교', text='소프트웨어마이스터고등학교'),
-                URIAction(label='홈페이지', uri='https://www.pycon.kr/')
-            ]
-        )
-    )
